Version 0.0.2 Update
Supported Devices:
AMAZON (Eclipse IDE)
ANDROID (Eclipse IDE)
IOS (XCODE IDE)
OSX - MAC OS X (XCODE IDE)
WIN32 - WINDOWS 32 bit System (Visual Studio 2013)
WINRT - WINDOWS METRO [Mobile Devices] (Visual Studio 2013)
Changes:
Added: WINRT - WINDOWS METRO [Mobile Devices] (Visual Studio 2013)
Added: Touch Screen Support (Both Mobile Devices and Desktop support)
Added: Accelerometer
Added: Joystick buttons
Updated: applovin sdk
Updated: admob sdk
Updated: rebmob sdk
Updated: improved support for winrt, win32, exe
Updated: Visual Studio 2012 to Visual Studio 2013
Fixed: applying banner on top of screen
Fixed: game play sounds in file not playing
Fixed: OSX game in full screen keyboard does not work
Fixed: one way collision
Fixed: do not background music is it already in play
Fixed: mute music/sound when displaying a video ad (Vungle)
Fixed: Connected Enemy Objects Disappear when destroyed with Bullet but
still collide and kill character
Fixed: do not stop death sound after death animation end
Fixed: Crashing intermittently on device after death
Fixed: Flat on flat collision with character walking will stick
eventually
Fixed: Downward bounce pre-set is broken
Fixed: admob and iAd banner issues
Fixed: chart-boost (and leadbolt) issue, game freezes
Fixed: game crashing at game over screen when pressing restart
Fixed: font size for Windows
Fixed: APK crashes on game over when not entering leader board ID
Fixed: reset on OS X does not reset progress
Fixed: character and background layering issues
Fixed: Pause Menu Back to Main Menu to Start Game x2 Crashes on Device
Fixed: Coin Shop functions differently when accessed from Main Menu
Fixed: Kill All Enemies Not Destroying Spawned Enemies
Fixed: sound for object/enemy does not play completely after second
appearance of object
Fixed: Start/End sounds not working correctly with coins, play sound on
Game Over Screen
Fixed: review button crashes when using “market://” URL on Google Play
Fixed: Label off alignment issue and add couple new scripts for building
CORE lib
Fixed: stop idle sound after Start Sound activated
Fixed: Power Up Sound Not Working (IDLE and END)
Fixed: Trail Effect Connected to Character Does Not Disappear When
Character Dies
Fixed: Back to Main Menu with Accelerometer causes crash
Fixed: OS X 10.10 bug with dialogue windows
Fixed: scroll bar for OSX 10.10 (auto hide scroll bar)
Fixed: object idle sound is not working
Fixed: score is not getting reported to leader board on game end scene
(only game over screen)
Fixed: Buttons from Main Menu can still be pressed in Shop despite not
being visible
Fixed: Power Up Sound Not Working (IDLE and END)
Fixed: added IDLE sound increase while character is closer to object

+++ ios/libs/cocos2dx/include/ccTypeInfo.h
@@ -41,7 +41,7 @@ public:
 
 static inline unsigned int getHashCodeByString(const char *key)
 {
-	unsigned int len = strlen(key);
+	unsigned long len = strlen(key);
 	const char *end=key+len;
 	unsigned int hash;
